查缺补漏

+++ UserServe/src/main/java/com/AAZl3l4/UserServe/controller/AddressController.java
@@ -6,13 +6,10 @@ import com.AAZl3l4.UserServe.service.IAddressService;
 import com.AAZl3l4.common.utils.Result;
 import com.AAZl3l4.common.utils.UserTool;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import io.seata.spring.annotation.GlobalTransactional;
 import io.swagger.v3.oas.annotations.Operation;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 
 @RestController
@@ -22,12 +19,19 @@ public class AddressController {
     @Autowired
     private IAddressService addressService;
 
+    @GetMapping("/list")
+    @Operation(summary = "获取用户地址")
+    public Result list() {
+        return Result.succeed(addressService.list(new QueryWrapper<Address>().eq("user_id", UserTool.getid())));
+    }
+
     @PostMapping("/add")
     @Operation(summary = "添加地址")
+    @GlobalTransactional(rollbackFor = Exception.class)
     public Result add(@RequestBody Address address) {
-        address.setId(UserTool.getid());
+        address.setUserId(UserTool.getid());
         if (address.getIsDefault()==('1')){
-            Address byId = addressService.getById(new QueryWrapper<Address>().eq("userId", address.getUserId()).eq("isDefault", '1'));
+            Address byId = addressService.getOne(new QueryWrapper<Address>().eq("user_id", address.getUserId()).eq("is_default", '1'));
             if (byId != null) {
                 byId.setIsDefault('0');
                 addressService.updateById(byId);
@@ -46,7 +50,7 @@ public class AddressController {
     @Operation(summary = "修改地址")
     public Result update(@RequestBody Address address) {
         if (address.getIsDefault()==('1')){
-            Address byId = addressService.getById(new QueryWrapper<Address>().eq("userId", address.getUserId()).eq("isDefault", '1'));
+            Address byId = addressService.getOne(new QueryWrapper<Address>().eq("user_id", address.getUserId()).eq("is_default", '1'));
             if (byId != null) {
                 byId.setIsDefault('0');
                 addressService.updateById(byId);
@@ -61,9 +65,12 @@ public class AddressController {
         }
     }
 
-    @PostMapping("/delete")
+    @PostMapping("/delete/{id}")
     @Operation(summary = "删除地址")
-    public Result delete(Integer id) {
+    public Result delete(@PathVariable("id") Integer id) {
+        if (!addressService.getById(id).getUserId().equals(UserTool.getid())){
+            return Result.error("没有权限");
+        }
         boolean delete = addressService.removeById(id);
 
         if (delete) {
